Changed learner overview and moved chart colors into one chart

+++ src/Education.Client/Features/History/Clients/Learner/Model/Examination/NumberOfExaminationsModel.cs
@@ -1,3 +1,3 @@
 namespace Education.Client.Features.History.Clients.Learner.Model.Examination;
 
-public sealed record NumberOfExaminationsModel(uint Succeeded, uint Failed, uint TimeExceeded, ulong Total);
+public sealed record NumberOfExaminationsModel(uint Succeeded, uint Failed, uint TimeExceeded, uint Total);
